utils: Move ResultExt to bootc_utils
This way it can be more easily used elsewhere.
