+++ crates/ostree-ext/src/container/store.rs
@@ -11,8 +11,8 @@ use crate::generic_decompress::Decompressor;
 use crate::logging::system_repo_journal_print;
 use crate::refescape;
 use crate::sysroot::SysrootLock;
-use crate::utils::ResultExt;
 use anyhow::{anyhow, Context};
+use bootc_utils::ResultExt;
 use camino::{Utf8Path, Utf8PathBuf};
 use canon_json::CanonJsonSerialize;
 use cap_std_ext::cap_std;
